local commit.

+++ spring-boot-hibernate-thymeleaf/src/main/java/com/spring/empDemo/repository/DepartmentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.spring.empDemo.model.entity.CityEntity;
 import com.spring.empDemo.model.entity.DepartmentEntity;
 
-public interface DepartmentRepository extends JpaRepository<DepartmentEntity, Integer>{
+public interface DepartmentRepository extends JpaRepository<DepartmentEntity, Long>{
 }
